Small fixes so the compilation works.

+++ src/main/java/com/street_art_explorer/resource_server/projection/UserRankingRow.java
@@ -9,9 +9,9 @@ public interface UserRankingRow {
 
     Integer getAuthServerUserId();
 
-    Long getMarkersCreated();
+    Integer getMarkersCreated();
 
-    Long getRatingsGiven();
+    Integer getRatingsGiven();
 
     Double getScore();
 }
